Refactor production tests
