+++ package.json
@@ -115,7 +115,8 @@
     "prepublishOnly": "echo \"Error: must publish from dist/\" && exit 1",
     "prepare-release": "yarn && yarn build && yarn test:dist",
     "test": "jest",
-    "test:dist": "node ./scripts/test-dist.js",
+    "test:dist": "cross-env NODE_ENV=production jest",
+    "test:dist-standalone": "cross-env NODE_ENV=production STANDALONE=1 jest tests/",
     "test:integration": "jest tests_integration",
     "perf:repeat": "yarn && yarn build && cross-env NODE_ENV=production node ./dist/bin-prettier.js --debug-repeat ${PERF_REPEAT:-1000} --loglevel debug ${PERF_FILE:-./index.js} > /dev/null",
     "perf:repeat-inspect": "yarn && yarn build && cross-env NODE_ENV=production node --inspect-brk ./dist/bin-prettier.js --debug-repeat ${PERF_REPEAT:-1000} --loglevel debug ${PERF_FILE:-./index.js} > /dev/null",
